testBiblio/requirements.txt
cryptography~=2.8
django~=2.2.7
djangorestframework~=3.12.2
django-cors-headers~=3.1.1
drf-jwt~=1.13.3
pyjwt~=1.7.1
requests~=2.22.0